+++ apps/frontend/app/routes/actions.tsx
@@ -1,7 +1,6 @@
 import { $path } from "@ignisda/remix-routes";
 import {
 	type ActionFunctionArgs,
-	type LoaderFunctionArgs,
 	json,
 	redirect,
 	unstable_parseMultipartFormData,
@@ -10,7 +9,6 @@ import {
 	AddEntityToCollectionDocument,
 	CommitMetadataDocument,
 	CommitPersonDocument,
-	CoreDetailsDocument,
 	CreateMediaReminderDocument,
 	CreateReviewCommentDocument,
 	DeleteMediaReminderDocument,
@@ -23,12 +21,8 @@ import {
 	RemoveEntityFromCollectionDocument,
 	ToggleMediaMonitorDocument,
 	ToggleMediaOwnershipDocument,
-	UserCollectionsListDocument,
-	UserDetailsDocument,
-	UserPreferencesDocument,
 	Visibility,
 } from "@ryot/generated/graphql/backend/graphql";
-import { safeRedirect } from "remix-utils/safe-redirect";
 import invariant from "tiny-invariant";
 import { match } from "ts-pattern";
 import { z } from "zod";
@@ -37,76 +31,15 @@ import { redirectToQueryParam } from "~/lib/generals";
 import {
 	MetadataSpecificsSchema,
 	colorSchemeCookie,
-	combineHeaders,
-	coreDetailsCookie,
 	createToastHeaders,
 	getAuthorizationHeader,
 	getLogoutCookies,
 	gqlClient,
 	processSubmission,
-	redirectIfNotAuthenticated,
 	s3FileUploader,
-	userCollectionsListCookie,
-	userDetailsCookie,
-	userPreferencesCookie,
 } from "~/lib/utilities.server";
 
-export const loader = async ({ request }: LoaderFunctionArgs) => {
-	await redirectIfNotAuthenticated(request);
-	const url = new URL(request.url);
-	const [
-		{ coreDetails },
-		{ userPreferences },
-		{ userDetails },
-		{ userCollectionsList },
-	] = await Promise.all([
-		gqlClient.request(CoreDetailsDocument),
-		gqlClient.request(
-			UserPreferencesDocument,
-			undefined,
-			await getAuthorizationHeader(request),
-		),
-		gqlClient.request(
-			UserDetailsDocument,
-			undefined,
-			await getAuthorizationHeader(request),
-		),
-		gqlClient.request(
-			UserCollectionsListDocument,
-			{},
-			await getAuthorizationHeader(request),
-		),
-	]);
-	const cookieMaxAge = coreDetails.tokenValidForDays * 24 * 60 * 60;
-	const redirectUrl = safeRedirect(
-		url.searchParams.get(redirectToQueryParam) || "/",
-	);
-	return redirect(redirectUrl, {
-		headers: combineHeaders(
-			{
-				"Set-Cookie": await coreDetailsCookie.serialize(coreDetails, {
-					maxAge: cookieMaxAge,
-				}),
-			},
-			{
-				"Set-Cookie": await userPreferencesCookie.serialize(userPreferences, {
-					maxAge: cookieMaxAge,
-				}),
-			},
-			{
-				"Set-Cookie": await userDetailsCookie.serialize(userDetails, {
-					maxAge: cookieMaxAge,
-				}),
-			},
-			{
-				"Set-Cookie": await userCollectionsListCookie.serialize(
-					userCollectionsList,
-					{ maxAge: cookieMaxAge },
-				),
-			},
-		),
-	});
-};
+export const loader = async () => redirect($path("/"));
 
 export const action = async ({ request }: ActionFunctionArgs) => {
 	const formData = await request.clone().formData();